refactor: mapeditorpage 책임 분리 (#25)
