+++ src/pages/MapEditorPage/components/ConnectionPanel.jsx
@@ -1,0 +1,518 @@
+import styled from 'styled-components'
+import Button from '../../../components/Button/Button'
+
+const SectionCard = styled.div`
+    border: 1px solid rgba(148, 163, 184, 0.18);
+    background: #f8fafc;
+    border-radius: 18px;
+    padding: 16px;
+    display: flex;
+    flex-direction: column;
+    gap: 14px;
+`
+
+const SectionTitle = styled.div`
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    gap: 12px;
+
+    h3 {
+        margin: 0;
+        font-size: 15px;
+        color: var(--black-900);
+    }
+
+    span {
+        font-size: 11px;
+        font-weight: 700;
+        color: var(--gray-500);
+    }
+`
+
+const HelpText = styled.div`
+    color: var(--gray-500);
+    font-size: 12px;
+    line-height: 1.6;
+`
+
+const ErrorHint = styled.div`
+    color: var(--red-500);
+    font-size: 11px;
+    line-height: 1.5;
+    margin-top: 4px;
+    text-align: left;
+`
+
+const DetailActions = styled.div`
+    display: flex;
+    gap: 8px;
+    flex-wrap: wrap;
+`
+
+const MappingGrid = styled.div`
+    display: grid;
+    grid-template-columns: repeat(auto-fit, minmax(240px, 1fr));
+    gap: 12px;
+`
+
+const MappingRow = styled.div`
+    border: 1px solid ${({ $active }) => ($active ? 'rgba(59, 130, 246, 0.32)' : 'rgba(148, 163, 184, 0.18)')};
+    background: ${({ $active }) => ($active ? 'rgba(239, 246, 255, 0.92)' : 'rgba(255, 255, 255, 0.96)')};
+    border-radius: 14px;
+    padding: 12px;
+    display: flex;
+    flex-direction: column;
+    gap: 10px;
+    box-shadow: ${({ $active }) => ($active ? '0 18px 32px -24px rgba(37, 99, 235, 0.45)' : 'none')};
+`
+
+const MappingRowHeader = styled.div`
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    gap: 12px;
+
+    strong {
+        font-size: 14px;
+        color: #0f172a;
+    }
+`
+
+const MappingStatus = styled.span`
+    font-size: 11px;
+    font-weight: 800;
+    color: ${({ $mapped }) => ($mapped ? '#047857' : '#b45309')};
+    background: ${({ $mapped }) => ($mapped ? 'rgba(16, 185, 129, 0.12)' : 'rgba(245, 158, 11, 0.12)')};
+    border-radius: 999px;
+    padding: 6px 8px;
+`
+
+const MappingMeta = styled.div`
+    color: #64748b;
+    font-size: 12px;
+    line-height: 1.5;
+`
+
+const ConnectionOverviewGrid = styled.div`
+    display: grid;
+    grid-template-columns: repeat(auto-fit, minmax(140px, 1fr));
+    gap: 10px;
+`
+
+const ConnectionStatCard = styled.div`
+    border: 1px solid rgba(148, 163, 184, 0.18);
+    background: linear-gradient(180deg, rgba(248, 250, 252, 0.96) 0%, rgba(255, 255, 255, 0.98) 100%);
+    border-radius: 16px;
+    padding: 14px;
+    display: flex;
+    flex-direction: column;
+    gap: 6px;
+`
+
+const ConnectionStatValue = styled.strong`
+    font-size: 24px;
+    line-height: 1;
+    color: #0f172a;
+`
+
+const ConnectionStatLabel = styled.span`
+    font-size: 12px;
+    font-weight: 700;
+    color: #64748b;
+`
+
+const ConnectionEmptyState = styled.div`
+    border: 1px dashed rgba(148, 163, 184, 0.28);
+    background: rgba(248, 250, 252, 0.72);
+    border-radius: 16px;
+    padding: 18px;
+    display: flex;
+    flex-direction: column;
+    gap: 10px;
+
+    strong {
+        font-size: 14px;
+        color: #0f172a;
+    }
+`
+
+const ConnectionPickBanner = styled.div`
+    border: 1px solid rgba(59, 130, 246, 0.22);
+    background: linear-gradient(180deg, rgba(239, 246, 255, 0.96) 0%, rgba(255, 255, 255, 0.98) 100%);
+    border-radius: 16px;
+    padding: 14px 16px;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    gap: 12px;
+    flex-wrap: wrap;
+`
+
+const ConnectionPickText = styled.div`
+    display: flex;
+    flex-direction: column;
+    gap: 4px;
+
+    strong {
+        font-size: 14px;
+        color: #0f172a;
+    }
+`
+
+const EntityList = styled.div`
+    display: flex;
+    flex-direction: column;
+    gap: 8px;
+    overflow: auto;
+    min-height: 0;
+`
+
+const EntityItem = styled.button`
+    width: 100%;
+    border: 1px solid ${({ $active }) => ($active ? 'rgba(79, 70, 229, 0.24)' : 'rgba(148, 163, 184, 0.16)')};
+    background: ${({ $active }) => ($active ? 'rgba(79, 70, 229, 0.12)' : 'rgba(255,255,255,0.95)')};
+    border-radius: 14px;
+    padding: 12px 14px;
+    cursor: pointer;
+    text-align: left;
+    display: flex;
+    flex-direction: column;
+    gap: 6px;
+    box-shadow: ${({ $active }) => ($active ? '0 18px 32px -24px rgba(91, 33, 182, 0.55)' : 'none')};
+    transform: ${({ $active }) => ($active ? 'translateY(-1px)' : 'none')};
+    transition: background 120ms ease, border-color 120ms ease, box-shadow 120ms ease, transform 120ms ease;
+`
+
+const EntityTitle = styled.div`
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    gap: 12px;
+
+    strong {
+        font-size: 14px;
+        font-weight: 800;
+        color: var(--black-900);
+    }
+
+    span {
+        font-size: 10px;
+        font-weight: 800;
+        color: #6366f1;
+        text-transform: uppercase;
+        letter-spacing: 0.04em;
+    }
+`
+
+const EntityMeta = styled.div`
+    color: var(--gray-600);
+    font-size: 12px;
+    line-height: 1.5;
+`
+
+const VerticalConnectorList = styled.div`
+    margin-top: 12px;
+    display: flex;
+    flex-direction: column;
+    gap: 8px;
+`
+
+const VerticalFloorRowsContainer = styled.div`
+    display: flex;
+    flex-direction: column;
+    gap: 6px;
+    margin-top: 8px;
+`
+
+const VerticalFloorRow = styled.div`
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    font-size: 13px;
+    background: #f8fafc;
+    padding: 8px 12px;
+    border-radius: 8px;
+    border: 1px solid #f1f5f9;
+`
+
+const VerticalFloorName = styled.span`
+    font-weight: 700;
+    color: #475569;
+    min-width: 48px;
+`
+
+const VerticalNodeStatus = styled.span`
+    margin-left: 12px;
+    color: ${({ $mapped }) => ($mapped ? '#0f172a' : '#94a3b8')};
+    flex-grow: 1;
+    font-weight: ${({ $mapped }) => ($mapped ? '600' : 'normal')};
+`
+
+const VerticalFloorRowActions = styled.div`
+    display: flex;
+    gap: 6px;
+`
+
+const DeleteTextButton = styled.button`
+    border: none;
+    background: none;
+    color: #ef4444;
+    font-size: 12px;
+    font-weight: 700;
+    cursor: pointer;
+    padding: 4px 8px;
+    display: inline-flex;
+    align-items: center;
+    border-radius: 4px;
+
+    &:hover {
+        background: rgba(239, 68, 68, 0.08);
+    }
+`
+
+export default function ConnectionPanel({
+    mappedGateCount,
+    campusGates,
+    buildingEntranceNodeCount,
+    verticalConnectionNodes,
+    pendingGatePick,
+    onCancelPendingGatePick,
+    entranceMappingByGateId,
+    floorNameById,
+    onStartGatePick,
+    isMappingLoading,
+    onFocusMappedNode,
+    entranceNodes,
+    entranceMappingByNodeId,
+    selectedEntity,
+    onSelectNode,
+    onOpenNodeTab,
+    verticalConnectors,
+    onCreateVerticalConnector,
+    isVerticalLoading,
+    activeConnectorForMapping,
+    floorOptions,
+    onDeleteVerticalConnector,
+    onViewVerticalNode,
+    onUnmapVerticalNode,
+    onStartVerticalNodePick,
+}) {
+    return (
+        <>
+            <SectionCard>
+                <SectionTitle>
+                    <h3>연결 현황</h3>
+                    <span>{mappedGateCount}/{campusGates.length} gate 연결</span>
+                </SectionTitle>
+                <ConnectionOverviewGrid>
+                    <ConnectionStatCard>
+                        <ConnectionStatValue>{campusGates.length}</ConnectionStatValue>
+                        <ConnectionStatLabel>캠퍼스 gate</ConnectionStatLabel>
+                    </ConnectionStatCard>
+                    <ConnectionStatCard>
+                        <ConnectionStatValue>{mappedGateCount}</ConnectionStatValue>
+                        <ConnectionStatLabel>연결 완료 gate</ConnectionStatLabel>
+                    </ConnectionStatCard>
+                    <ConnectionStatCard>
+                        <ConnectionStatValue>{buildingEntranceNodeCount}</ConnectionStatValue>
+                        <ConnectionStatLabel>건물 전체 entrance node</ConnectionStatLabel>
+                    </ConnectionStatCard>
+                    <ConnectionStatCard>
+                        <ConnectionStatValue>{verticalConnectionNodes.length}</ConnectionStatValue>
+                        <ConnectionStatLabel>수직 이동 후보</ConnectionStatLabel>
+                    </ConnectionStatCard>
+                </ConnectionOverviewGrid>
+                <HelpText>
+                    gate 매핑은 건물 단위로 한 번만 관리합니다. 현재 층에 entrance node가 없어도, 다른 층에서 이미 연결된 gate는 여기서 그대로 연결 완료로 보여야 합니다.
+                </HelpText>
+            </SectionCard>
+
+            <SectionCard>
+                <SectionTitle>
+                    <h3>출입구 캘리브레이션</h3>
+                    <span>{campusGates.length}개 gate</span>
+                </SectionTitle>
+                <HelpText>
+                    Gate를 하나 고른 뒤 지도에서 노드를 직접 클릭해 연결합니다. 이미 연결된 gate도 언제든 다시 지정할 수 있습니다.
+                </HelpText>
+                {pendingGatePick && (
+                    <ConnectionPickBanner>
+                        <ConnectionPickText>
+                            <strong>{pendingGatePick.gateName}에 연결할 노드를 지도에서 클릭하세요.</strong>
+                            <HelpText>아무 node나 지정할 수 있고, 연결되면 해당 node는 자동으로 entrance로 처리됩니다.</HelpText>
+                        </ConnectionPickText>
+                        <Button variant="outlineGray" size="sm" onClick={onCancelPendingGatePick}>
+                            선택 취소
+                        </Button>
+                    </ConnectionPickBanner>
+                )}
+                {campusGates.length > 0 ? (
+                    <MappingGrid>
+                        {campusGates.map((gate) => {
+                            const mapped = entranceMappingByGateId.get(gate.id)
+                            return (
+                                <MappingRow key={gate.id} $active={pendingGatePick?.gateId === gate.id}>
+                                    <MappingRowHeader>
+                                        <strong>{gate.name}</strong>
+                                        <MappingStatus $mapped={!!mapped}>
+                                            {pendingGatePick?.gateId === gate.id ? '노드 선택 중' : mapped ? '연결 완료' : '미연결'}
+                                        </MappingStatus>
+                                    </MappingRowHeader>
+                                    <MappingMeta>
+                                        {mapped?.nodeId
+                                            ? `${gate.name}이 ${floorNameById.get(mapped.floorId) || '다른 층'}의 실내 node와 연결되어 있습니다.`
+                                            : '아직 연결된 실내 node가 없습니다. 지도에서 바로 지정해 주세요.'}
+                                    </MappingMeta>
+                                    <DetailActions>
+                                        <Button size="sm" onClick={() => onStartGatePick(gate)} disabled={isMappingLoading}>
+                                            {mapped?.nodeId ? '다시 지정' : '지도에서 지정'}
+                                        </Button>
+                                        {mapped?.nodeId && (
+                                            <Button variant="outlineGray" size="sm" onClick={() => onFocusMappedNode(mapped.nodeId)}>
+                                                지도에서 보기
+                                            </Button>
+                                        )}
+                                        {pendingGatePick?.gateId === gate.id && (
+                                            <Button variant="outlineGray" size="sm" onClick={onCancelPendingGatePick}>
+                                                취소
+                                            </Button>
+                                        )}
+                                    </DetailActions>
+                                </MappingRow>
+                            )
+                        })}
+                    </MappingGrid>
+                ) : (
+                    <ErrorHint>캠퍼스에 등록된 gate가 없습니다. 캠퍼스 관리에서 gate를 먼저 등록해 주세요.</ErrorHint>
+                )}
+            </SectionCard>
+
+            <SectionCard>
+                <SectionTitle>
+                    <h3>실내 출입구 노드</h3>
+                    <span>{entranceNodes.length}개</span>
+                </SectionTitle>
+                <HelpText>
+                    현재 층의 entrance node 목록입니다. 건물 전체 연결 여부와는 별개로, 이 층에서 바로 조정할 수 있는 node만 보여줍니다.
+                </HelpText>
+                {entranceNodes.length > 0 ? (
+                    <EntityList>
+                        {entranceNodes.map((node) => {
+                            const mapping = entranceMappingByNodeId.get(node.id)
+                            const active = selectedEntity?.type === 'node' && selectedEntity.id === node.id
+                            return (
+                                <EntityItem key={node.id} type="button" $active={active} onClick={() => onSelectNode(node.id)}>
+                                    <EntityTitle>
+                                        <strong>{node.name || '이름 없는 출입구'}</strong>
+                                        <span>{mapping?.campusGateName || 'UNMAPPED'}</span>
+                                    </EntityTitle>
+                                    <EntityMeta>
+                                        {mapping?.campusGateName ? `${mapping.campusGateName}와 연결됨` : '아직 gate와 연결되지 않았습니다.'}
+                                    </EntityMeta>
+                                </EntityItem>
+                            )
+                        })}
+                    </EntityList>
+                ) : (
+                    <ConnectionEmptyState>
+                        <strong>현재 층에 entrance node가 없습니다.</strong>
+                        <HelpText>
+                            다른 층에서 이미 gate 매핑이 끝났을 수도 있습니다. 이 영역은 현재 층에서 직접 수정할 entrance node가 있을 때만 표시됩니다.
+                        </HelpText>
+                        <DetailActions>
+                            <Button size="sm" onClick={onOpenNodeTab}>
+                                노드 탭 열기
+                            </Button>
+                        </DetailActions>
+                    </ConnectionEmptyState>
+                )}
+            </SectionCard>
+
+            <SectionCard>
+                <SectionTitle>
+                    <h3>수직 이동 연결 (캘리브레이션)</h3>
+                    <Button size="sm" onClick={onCreateVerticalConnector} disabled={isVerticalLoading}>
+                        + 추가
+                    </Button>
+                </SectionTitle>
+                <HelpText>
+                    엘리베이터, 계단, 에스컬레이터 등 수직 통로를 만들고 각 층의 노드를 연결합니다. 동일한 통로에 연결된 노드들 간에 자동으로 수직 이동 가중치가 생성됩니다.
+                </HelpText>
+
+                {verticalConnectors.length > 0 ? (
+                    <VerticalConnectorList>
+                        {verticalConnectors.map((connector) => {
+                            const icon = connector.kind === 'elevator' ? '🛗' : connector.kind === 'stair' ? '🪜' : connector.kind === 'escalator' ? '🛗' : '🪜'
+                            const isActive = activeConnectorForMapping?.connectorId === connector.id
+                            return (
+                                <MappingRow key={connector.id} $active={isActive}>
+                                    <MappingRowHeader>
+                                        <strong>{icon} {connector.name}</strong>
+                                        <DeleteTextButton type="button" onClick={() => onDeleteVerticalConnector(connector.id)} disabled={isVerticalLoading}>
+                                            삭제
+                                        </DeleteTextButton>
+                                    </MappingRowHeader>
+                                    <MappingMeta>
+                                        종류: {connector.kind} ({connector.nodes?.length || 0}개 층 연결됨)
+                                    </MappingMeta>
+                                    <VerticalFloorRowsContainer>
+                                        {floorOptions.map((floor) => {
+                                            const mappedNode = connector.nodes?.find((node) => node.floorId === floor.id)
+                                            const isCurrentPicking =
+                                                activeConnectorForMapping?.connectorId === connector.id &&
+                                                activeConnectorForMapping?.floorId === floor.id
+
+                                            return (
+                                                <VerticalFloorRow key={floor.id}>
+                                                    <VerticalFloorName>{floor.name}</VerticalFloorName>
+                                                    <VerticalNodeStatus $mapped={!!mappedNode}>
+                                                        {isCurrentPicking
+                                                            ? '지도에서 노드 선택 중...'
+                                                            : mappedNode
+                                                                ? `${mappedNode.nodeName || '연결됨'} (${mappedNode.nodeId?.substring(0, 8)})`
+                                                                : '미연결'}
+                                                    </VerticalNodeStatus>
+                                                    <VerticalFloorRowActions>
+                                                        {mappedNode ? (
+                                                            <>
+                                                                <Button variant="outlineGray" size="sm" onClick={() => onViewVerticalNode(floor.id, mappedNode.nodeId)}>
+                                                                    보기
+                                                                </Button>
+                                                                <Button
+                                                                    variant="dangerOutline"
+                                                                    size="sm"
+                                                                    onClick={() => onUnmapVerticalNode(connector.id, floor.id)}
+                                                                    disabled={isVerticalLoading}
+                                                                >
+                                                                    해제
+                                                                </Button>
+                                                            </>
+                                                        ) : (
+                                                            <Button
+                                                                size="sm"
+                                                                onClick={() => onStartVerticalNodePick(connector.id, connector.name, floor.id)}
+                                                                disabled={isVerticalLoading || isCurrentPicking}
+                                                            >
+                                                                {isCurrentPicking ? '선택 중' : '지정'}
+                                                            </Button>
+                                                        )}
+                                                    </VerticalFloorRowActions>
+                                                </VerticalFloorRow>
+                                            )
+                                        })}
+                                    </VerticalFloorRowsContainer>
+                                </MappingRow>
+                            )
+                        })}
+                    </VerticalConnectorList>
+                ) : (
+                    <ConnectionEmptyState>
+                        <strong>등록된 수직 이동수단이 없습니다.</strong>
+                        <HelpText>우측 상단의 "+ 추가" 버튼을 눌러 엘리베이터나 계단을 먼저 만들어보세요.</HelpText>
+                    </ConnectionEmptyState>
+                )}
+            </SectionCard>
+        </>
+    )
+}
